+++ packages/fiori-app-sub-generator/test/int/fiori-freestyle/expected-output/simple_v2_mock_tasks/webapp/test/integration/opaTests.qunit.js
@@ -1,0 +1,7 @@
+/* global QUnit */
+
+sap.ui.require(["simplev2mocktasks/test/integration/AllJourneys"
+], function () {
+	QUnit.config.autostart = false;
+	QUnit.start();
+});